Code/week4/requirements.txt
# Core
torch>=2.2.0
torchvision>=0.17.0
numpy
tqdm
pillow

# Verification
z3-solver>=4.12.2

# ONNX export
onnx
onnxruntime

# Adversarial / analysis
matplotlib
pandas
scikit-learn

torchattacks
